ibv: eight-byte align "packed long" payloads
This commit introduces eight-byte alignment for "packed long"
payloads.  Previously, a sufficiently small Long payload would be
packed immediately following the last 32-bit argument.  That means
that on LP64 ABIs, even argument counts would yield a payload which
was four-byte aligned but never eight-byte aligned.

Isolated `memcpy` benchmarks show the lack of eight-byte alignment can
yield a 4% to 14% reduction in bandwidth on Dirac.

On Summit, `testam` results show measurable improvements in bandwidth
for the "AMLong ping-pong roundtrip ReqRep" sub-test from 128 bytes
(one cache line) through 2048 (the largest power of two using packed
long with default settings).  Below 128 bytes the data is too noisy to
make any before/after distinction.  At 2048-byte payload, this small
change yields a mean (over 10 runs) bandwidth improvement of 3.1%, as
compared to a relative standard deviation of 0.83%.  At 1024 bytes,
the bandwidth improvement is 1.4%, vs a relative std dev of 0.56%.

In addition to the performance benefits noted above, this change will
help an upcoming native implementation of NPAM Long to provide a
buffer "suitably aligned to hold any data type" while leveraging the
existing packed long logic for appropriately sized payloads.

+++ ibv-conduit/gasnet_core.c
@@ -1266,7 +1266,7 @@ static int gasnetc_load_settings(void) {
   if_pf (gasnetc_packedlong_limit > GASNETC_MAX_PACKEDLONG) {
     fprintf(stderr,
             "WARNING: GASNET_PACKEDLONG_LIMIT reduced from %u to %u\n",
-            (unsigned int)gasnetc_packedlong_limit, GASNETC_MAX_PACKEDLONG);
+            (unsigned int)gasnetc_packedlong_limit, (unsigned int)GASNETC_MAX_PACKEDLONG);
     gasnetc_packedlong_limit = GASNETC_MAX_PACKEDLONG;
   }
